Emit error rather than crash
In this instance, 'dps' is not a valid key in status, so the next line will crash.  For my device, at least, status is a well formed json error packet that matches the structure produced by error_json, so returning it seems reasonable.  My one caveat is that I can't get my device to work, so I can't be certain that a successful call will not also contain a key called Error.  If it does, you can change the code to test for the absence of dps.